pyproject.toml
[project]
name = "cryptohedge"
version = "0.1.0"
description = "Autonomous 7-agent AI system for crypto trading"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai",
    "duckdb",
    "ccxt",
    "pandas-ta",
    "python-telegram-bot",
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "python-dotenv",
    "grpcio",
    "grpcio-tools",
    "rich",
    "click",
    "textual",
    "pydantic",
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[project.scripts]
cryptohedge = "src.cli.main:cli"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
